+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [2-beta18 - 2025-05-08]
+- updated deps (Tableplot, Clay)
+
+## [2-beta17 - 2025-04-30]
+- updated deps (Kindly, Clay, Kindly-advice)
+
+## [2-beta16 - 2025-04-18]
+- updated Clay version
+
 ## [2-beta15 - 2025-04-09]
 - updated Clay version